feat: gpu graph coloring
Based on Vivace paper. Performances are not great at this stage. Couple of optimization that could improve the situation:
- Use IndirectDispatch and let the previous kernel decide what should be the next dispatch size
- Push constraints to be colored in a buffer and dispatch only based on its size
- Run multiple run before checking if we converged.
- Use interleaved buffers to improve cache hit
- Replace Random generator by hashes (trade bandwidth against computation)

The bottleneck will be the neibourhood discovery. This could be further addressed by building it
while adding new constraints. This should be fairly static.

At this stage it perform as well as the CPU version :(